connected process_agreements

+++ agents/super_agent/utils/utils.py
@@ -2,4 +2,4 @@ from time import time
 
 
 def get_ms_current_time():
-    return time() * 1000
+    return int(time() * 1000)
